+++ changelog.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+### 9.17.1
+
+* `[strutil]` Added usage example for `Copy` method
+* `[system/procname]` Added usage examples
+
 ### 9.17.0
 
 * `[netutil]` Ignore TUN/TAP interfaces while searching main IP address